Daily notifications - Added cURL
This is super messy and I need to make it prettier when I have time - but this will make SLAP alert you every day rather than JUST when there's a difference. Main reason for this is just to be notified that SLAP is functioning correctly if you've cron job'd it.